backend/requirements.txt
# FastAPI and web
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
httpx==0.26.0

# AWS
aioboto3==12.3.0
boto3==1.34.34

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# Auth
python-jose[cryptography]==3.3.0

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.1

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
moto[all]>=5.0.0
aiosqlite>=0.19.0

# Development
black==24.1.1
ruff==0.2.0
mypy==1.8.0
